+++ src/main/java/com/parim/model/components/MarioObject.java
@@ -1,7 +1,7 @@
 package com.parim.model.components;
 
 public class MarioObject extends TileObject{
-    public MarioObject(int x, int y, String type) {
+    public MarioObject(double x, double y, String type) {
         super(x, y, type);
     }
 }
