+++ src/main/java/com/parim/model/components/TileObject.java
@@ -2,12 +2,12 @@ package com.parim.model.components;
 
 public abstract class TileObject {
     protected static int SIZE = 90;
-    protected int x, y;
-    protected int xVelocity, yVelocity;
+    protected double x, y;
+    protected double xVelocity, yVelocity;
     protected String type;
 
     public TileObject(){}
-    public TileObject(int x, int y, String type){
+    public TileObject(double x, double y, String type){
         this.x = x;
         this.y = y;
         this.type = type;
@@ -28,35 +28,35 @@ public abstract class TileObject {
         TileObject.SIZE = SIZE;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int getXVelocity() {
+    public double getXVelocity() {
         return xVelocity;
     }
 
-    public void setXVelocity(int xVelocity) {
+    public void setXVelocity(double xVelocity) {
         this.xVelocity = xVelocity;
     }
 
-    public int getYVelocity() {
+    public double getYVelocity() {
         return yVelocity;
     }
 
-    public void setYVelocity(int yVelocity) {
+    public void setYVelocity(double yVelocity) {
         this.yVelocity = yVelocity;
     }
